+++ modules/alloy/src/main/ts/ephox/alloy/ui/slider/ModelCommon.ts
@@ -1,26 +1,25 @@
-import { TouchEvent, MouseEvent, Touch } from '@ephox/dom-globals';
-import { Option, Fun } from '@ephox/katamari';
+import { MouseEvent, Touch, TouchEvent } from '@ephox/dom-globals';
+import { Fun, Option } from '@ephox/katamari';
 import { Position } from '@ephox/sugar';
-import { PlatformDetection } from '@ephox/sand';
 
 import { SugarPosition } from '../../alien/TypeDefinitions';
 import { NativeSimulatedEvent } from '../../events/SimulatedEvent';
 
-const isTouch = PlatformDetection.detect().deviceType.isTouch();
-
 const _sliderChangeEvent = 'slider.change.value';
 const sliderChangeEvent = Fun.constant(_sliderChangeEvent);
 
+const isTouchEvent = (evt: MouseEvent | TouchEvent): evt is TouchEvent => evt.type.indexOf('touch') !== -1;
+
 const getEventSource = (simulatedEvent: NativeSimulatedEvent): Option<SugarPosition> => {
   const evt = simulatedEvent.event().raw();
-  if (isTouch) {
-    const touchEvent = evt as TouchEvent;
+  if (isTouchEvent(evt)) {
+    const touchEvent = evt;
     return touchEvent.touches !== undefined && touchEvent.touches.length === 1 ?
       Option.some(touchEvent.touches[0]).map((t: Touch) => {
         return Position(t.clientX, t.clientY) as SugarPosition;
       }) : Option.none();
   } else {
-    const mouseEvent = evt as MouseEvent;
+    const mouseEvent = evt;
     return mouseEvent.clientX !== undefined ? Option.some(mouseEvent).map((me) => {
       return Position(me.clientX, me.clientY) as SugarPosition;
     }) : Option.none();