+++ modules/alloy/changelog.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Changed touch/mouse event handling to work with hybrid devices that accept both mouse and touch events.
+- Changed `AlloyEvents.runActionExtra()` to pass the simulated event to the callback.
+
+### Fixed
+- Fixed the `Slider` component not working in some cases on touch devices.
 
 # [5.0.7] - 2019-12-02